bdd: Move database functions to a test library called stores
This code is going to be used in the functional tests that will be
written while the code is built.
