+++ test/features/support/steps.js
@@ -1,14 +1,15 @@
-/* Test tools */
-import sinon from 'sinon';
 import Promise from 'bluebird';
 
+/* Test tools */
+import sinon from 'sinon';
 import { After, Before, Given, When, Then } from 'cucumber';
 import { expect } from 'chai';
 import { resetTestDB } from '../../utils';
+import * as libteststore from '../../lib/stores';
 
 /* Required for building context  */
-import models from '../../../server/models';
 import * as stripe from '../../../server/paymentProviders/stripe/gateway';
+import models from '../../../server/models';
 
 /* What's being tested */
 import * as libledger from '../../../server/lib/ledger';
@@ -16,13 +17,6 @@ import * as libpayments from '../../../server/lib/payments';
 
 /* Sandbox for mocks, stubs, etc.*/
 const sandbox = sinon.sandbox.create();
-
-export function randEmail(email) {
-  const [user, domain] = email.split('@');
-  const rand = Math.random().toString(36).substring(2, 15);
-  return `${user}-${rand}@${domain}`;
-}
-
 
 /* Setup environment for each test.
  *
@@ -35,41 +29,19 @@ Before(async () => {
   await resetTestDB();
 });
 
+/* Reset the mocks after each run */
 After(() => sandbox.restore());
 
 Given('a User {string}', async function (name) {
-  const email = randEmail(`${name}@oc.com`);
-  const user = await models.User.createUserWithCollective({
-    email,
-    name,
-    username: name,
-    description: `A user called ${name}`,
-  });
-  this.transaction.keys[name] = user.collective;
+  const { userCollective } = await libteststore.newUser(name);
+  this.transaction.keys[name] = userCollective;
 });
 
 Given(/^a Collective "([^\"]+)" with a host in "([^\"]+)"( and "([^\"]+)%" fee)?$/, async function (name, currency, fee) {
-  const email = randEmail(`${name}-host-${currency}@oc.com`);
-  const hostOwner = await models.User.create({ email });
-  const host = await models.Collective.create({
-    CreatedByUserId: hostOwner.id,
-    slug: "Host",
-    hostFeePercent: fee ? parseInt(fee) : 0,
-    currency,
-  });
-
-  const collective = await models.Collective.create({ name });
-  await collective.addHost(host);
-  await models.ConnectedAccount.create({
-    service: 'stripe',
-    token: 'sk_test_XOFJ9lGbErcK5akcfdYM1D7j',
-    username: 'acct_198T7jD8MNtzsDcg',
-    CollectiveId: host.id,
-  });
-
+  const { host, collective } = await libteststore.collectiveWithHost(name, currency, fee);
   this.transaction.keys[name] = collective;
   this.transaction.keys[`${name}-host`] = host;
-  this.transaction.set({ hostOwner, host, collective });
+  this.transaction.set({ host, collective });
 });
 
 Given('the conversion rate from {string} to {string} is {float}', function (from, to, rate) {
@@ -84,34 +56,25 @@ When('{string} donates {string} to {string}', async function (from, amount, to) 
   const [value, currency] = amount.split(' ');
   const fromNode = this.transaction.keys[from];
   const toNode = this.transaction.keys[to];
-  const order = await models.Order.create({
-    description: 'Donation',
-    totalAmount: parseInt(value),
-    currency,
-    CreatedByUserId: fromNode.CreatedByUserId,
-    FromCollectiveId: fromNode.id,
-    CollectiveId: toNode.id,
-  });
-  await order.setPaymentMethod({
-    token: "tok_123456781234567812345678",
-  });
+  const { order } = await libteststore.orderAndPaymentMethod(
+    fromNode, toNode, parseInt(value), currency);
 
+  /* This has to be called before executeOrder() and it's being called
+   * here and not in the hook because it needs the order parameters to
+   * be generated. */
   sandbox.stub(
     stripe,
     "retrieveBalanceTransaction",
     this.transaction.createRetrieveBalanceTransactionStub(value, currency));
 
-  const user = models.User.findById(fromNode.CreatedByUserId);
-
+  const user = await models.User.findById(fromNode.CreatedByUserId);
   const transaction = await libpayments.executeOrder(user, order);
-
   this.transaction.set({ order, transaction });
 });
 
 Then('{string} should have {string} in their balance', function(ledger, amount) {
   const ledgerId = this.transaction.keys[ledger].id;
-  const [value, currency] = amount.split(' ');
-  console.log(value, currency, ledgerId);
-
-  //expect(this.variable).to.eql(number);
+  const rows = libledger.rows(this.transaction.state.transaction.TransactionGroup);
+  const balance = libledger.balanceFromCurrency(ledgerId, rows);
+  // expect(balance).to.equal(amount);
 });
